Dragonexpert patch 1

+++ inc/plugins/socialgroups/classes/socialgroupsuserhandler.php
@@ -3,9 +3,10 @@
  * Socialgroups plugin created by Mark Janssen.
  * This is not a free plugin.
  */
+
 if(!defined("IN_MYBB"))
 {
-    die("Direct access not allowed.");
+    die("Direct access to the user handler is not allowed.");
 }
 
 class socialgroupsuserhandler
@@ -32,17 +33,20 @@ class socialgroupsuserhandler
         // Nothing
     }
 
-    /* This function checks if a person is a member of the specified group. */
+    /**
+     * This function checks if a person is a member of the specified group.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     * @return bool Whether the user is a member
+     */
 
-    public function is_member($gid=1, $uid=0)
+    public function is_member(int $gid=1, int $uid=0): bool
     {
         global $mybb, $db, $socialgroups;
-        $gid = (int) $gid;
         if(!$gid)
         {
             $socialgroups->error("invalid_group");
         }
-        $uid = (int) $uid;
         if(!$uid)
         {
             $uid = $mybb->user['uid'];
@@ -62,30 +66,31 @@ class socialgroupsuserhandler
         else
         {
             // we have to manually fetch
-            $query = $db->simple_select("socialgroup_members", "*", "gid=$gid");
-            while($user = $db->fetch_array($query))
+            $query = $db->simple_select("socialgroup_members", "*", "gid=" . $gid . " AND uid=" . $uid);
+            if($db->num_rows($query) != 0)
             {
-                $this->members[$gid][] = (int) $user['uid'];
-            }
-            if(in_array($uid, $this->members[$gid]))
-            {
+                $user = $db->fetch_array($query);
+                $this->members[$gid][] = $user['uid'];
                 return true;
             }
             return false;
         }
     }
 
-    /* This function checks if a person is a group leader. */
+    /**
+     * This function checks if a person is a group leader.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     * @return bool Whether the user is a leader of the group.
+     */
 
-    public function is_leader($gid=1, $uid=0)
+    public function is_leader(int $gid=1, int $uid=0): bool
     {
         global $db, $mybb, $socialgroups;
-        $gid = (int) $gid;
         if(!$gid)
         {
             $socialgroups->error("invalid_group");
         }
-        $uid = (int) $uid;
         if(!$uid)
         {
             $uid = $mybb->user['uid'];
@@ -107,6 +112,7 @@ class socialgroupsuserhandler
             {
                 $this->leaders[$gid][$leader['uid']] = $leader['uid'];
             }
+            $db->free_result($query);
             if(in_array($uid, $this->leaders[$gid]))
             {
                 return true;
@@ -115,9 +121,14 @@ class socialgroupsuserhandler
         }
     }
 
-    /* This function checks if a person is a moderator.*/
+    /**
+     * This function checks if a person is a moderator.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     * @return bool Whether the user is a moderator of the group.
+     */
 
-    public function is_moderator($gid=1, $uid=0)
+    public function is_moderator(int $gid=1, int $uid=0): bool
     {
         global $mybb, $db;
         // Admins and global moderators should be automatic moderators
@@ -125,13 +136,11 @@ class socialgroupsuserhandler
         {
             return true;
         }
-        $gid = (int) $gid;
         if(!$gid)
         {
             // We return false here instead of above because admins should be allowed to moderate anything.
             return false;
         }
-        $uid = (int) $uid;
         if(!$uid)
         {
             $uid = $mybb->user['uid'];
@@ -145,18 +154,21 @@ class socialgroupsuserhandler
         return false;
     }
 
-    /* This function loads the members of a group. */
+    /**
+     * This function loads the members of a group.
+     * @param int $gid The id of the group.
+     * @return Array An array of group members.
+     */
 
-    public function load_members($gid)
+    public function load_members(int $gid=0): array
     {
         global $db, $socialgroups;
-        $gid = (int) $gid;
         if(array_key_exists($gid, $this->members))
         {
             return $this->members[$gid];
         }
         $query = $db->simple_select("socialgroup_members", "*", "gid=$gid");
-        if($db->num_rows($query) == 0)
+        if(!$db->num_rows($query))
         {
             $socialgroups->error("invalid_group");
         }
@@ -164,14 +176,17 @@ class socialgroupsuserhandler
         {
             $this->members[$gid][] = (int) $user['uid'];
         }
+        $db->free_result($query);
         return $this->members[$gid];
     }
 
-    /* This function loads the moderators for social groups.  Typically called after fetching a group.
-    * @Param 1: an optional parameter used to cache data if you need to call it later in the script.
-    * @Return: associative array with the keys users and usergroups, both of which are arrays. */
+    /**
+     * This function loads the moderators for social groups.  Typically called after fetching a group.
+    * @param int $gid The id of the group.
+    * @return Array Associative array with the keys users and usergroups, both of which are arrays.
+     */
 
-    public function load_moderators($gid=1)
+    public function load_moderators(int $gid=1) : array
     {
         global $db;
         if(array_key_exists($gid, $this->moderators))
@@ -184,6 +199,7 @@ class socialgroupsuserhandler
         {
             $this->moderators[$gid]['users'][] = $moderator['uid'];
         }
+        $db->free_result($query);
         // Now capture the usergroups
         $query = $db->simple_select("usergroups", "gid", "issupermod=1");
         $this->moderators[$gid]['usergroups'] = array();
@@ -191,36 +207,47 @@ class socialgroupsuserhandler
         {
             $this->moderators[$gid]['usergroups'][] = $moderator['gid'];
         }
+        $db->free_result($query);
         return $this->moderators[$gid];
     }
 
-    public function load_leaders($gid=1)
+    /**
+     * This function loads all leaders of a group.
+     * @param int $gid The id of the group.
+     * @return Array An array of leaders.
+     */
+
+    public function load_leaders(int $gid=1): array
     {
         global $db;
         if(array_key_exists($gid, $this->leaders))
         {
             return $this->leaders[$gid];
         }
-        $query = $db->simple_select("socialgroup_leaders", "uid", "gid=" . (int) $gid);
+        $query = $db->simple_select("socialgroup_leaders", "uid", "gid=" . $gid);
         while($leader = $db->fetch_array($query))
         {
             $this->leaders[$gid][] = $leader['uid'];
         }
+        $db->free_result($query);
         return $this->leaders[$gid];
     }
 
-    /* This checks if a person is able to join a group. */
+    /**
+     * This checks if a person is able to join a group.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     * @return bool Whether the user can join.
+     */
 
-    public function can_join($gid=1, $uid=0)
+    public function can_join(int $gid=1, int $uid=0): bool
     {
-        global $mybb, $db;
-        $gid = (int) $gid;
+        global $mybb;
         if(!$gid)
         {
             return false;
         }
         $cid = $this->group[$gid]['cid'];
-        $uid = (int) $uid;
         if(!$uid)
         {
             $uid = $mybb->user['uid'];
@@ -252,13 +279,16 @@ class socialgroupsuserhandler
         return true;
     }
 
-    /* This function checks if a member has an invitation to join a group. */
+    /**
+     * This function checks if a member has an invitation to join a group.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     * @return bool Whether the user has an invite to the group.
+     */
 
-    public function has_invite($gid=1, $uid=0)
+    public function has_invite(int $gid=1, int $uid=0): bool
     {
         global $db, $mybb;
-        $gid = (int) $gid;
-        $uid = (int) $uid;
         if(!$gid)
         {
             return false;
@@ -279,7 +309,14 @@ class socialgroupsuserhandler
         return false;
     }
 
-    public function has_join_request($gid=0, $uid=0)
+    /**
+     * This checks if a join request exists for the user.
+     * @param int $gid  The id of the group.
+     * @param int $uid The id of the user.
+     * @return bool Whether the user has a join request for the group.
+     */
+
+    public function has_join_request(int $gid=0, int $uid=0): bool
     {
         global $db;
         $query = $db->simple_select("socialgroup_join_requests", "*", "gid=" . $gid . " AND uid=" . $uid);
@@ -292,14 +329,16 @@ class socialgroupsuserhandler
         return false;
     }
 
-    /* This function puts a member in a group.
-    * The force parameter forces a person into the group.  Useful for ACP implementation or invites. */
+    /**
+     * This function puts a member in a group.
+     * @param int $gid The id of the group.
+     * @param int $uid THe id of the user.
+     * @param int $force Whether to insert regardless of permissions.
+     */
 
-    public function join($gid=0, $uid=0, $force=0)
+    public function join(int $gid=0, int $uid=0, int $force=0)
     {
         global $db, $mybb, $socialgroups;
-        $gid = (int) $gid;
-        $uid = (int) $uid;
         if(!$uid)
         {
             $uid = $mybb->user['uid'];
@@ -310,16 +349,21 @@ class socialgroupsuserhandler
         }
         $new_member = array(
             "uid" => $uid,
-            "gid" => $gid
+            "gid" => $gid,
+            "dateline" => time()
         );
         $db->insert_query("socialgroup_members", $new_member);
         $db->delete_query("socialgroup_invites", "touid=$uid AND gid=$gid");
         $db->delete_query("socialgroup_join_requests", "uid=$uid AND gid=$gid");
     }
 
-    /* Remove a member from a group. */
+    /**
+     * Remove a member from a group.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     */
 
-    public function remove_member($gid, $uid=0)
+    public function remove_member(int $gid=0, int $uid=0)
     {
         global $mybb, $db, $socialgroups;
         if(!$this->is_member($gid, $uid) || $this->is_leader($gid, $uid))
@@ -333,10 +377,16 @@ class socialgroupsuserhandler
         $db->delete_query("socialgroup_members", "uid=$uid AND gid=$gid");
     }
 
-    public function add_leader($gid, $uid=0)
+    /**
+     * This function adds a leader to a group.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     * @return false|int False on failure and the id of the leader on success.
+     */
+
+    public function add_leader(int $gid=0, int $uid=0)
     {
         global $db, $mybb;
-        $uid = (int) $uid;
         if(!$uid)
         {
             $uid = $mybb->user['uid'];
@@ -347,14 +397,22 @@ class socialgroupsuserhandler
             return false;
         }
         $new_leader = array(
-            "gid" => (int) $gid,
-            "uid" => (int) $uid
+            "gid" =>  $gid,
+            "uid" =>  $uid
         );
         $lid = $db->insert_query("socialgroup_leaders", $new_leader);
         return $lid;
     }
 
-    public function remove_leader($gid, $uid=0, $admin_overide = false)
+    /**
+     * This removes a leader from a group.
+     * @param int $gid The id of the group.
+     * @param int $uid The id of the user.
+     * @param bool $admin_overide Allow an admin overide. Mainly for Admin CP.
+     * @return bool Whether the leader got deleted.
+     */
+
+    public function remove_leader(int $gid=0, int $uid=0, bool $admin_overide = false): bool
     {
         global $db, $mybb, $socialgroups;
         // First load the group.
@@ -397,9 +455,22 @@ class socialgroupsuserhandler
         }
     }
 
-    public function transfer_ownership($original_owner, $new_owner, $gid, $stay_leader=true)
+    /**
+     * This transfers ownership of the group to someone else.
+     * @param int $original_owner The id of the owner.
+     * @param int $new_owner The id of the new owner.
+     * @param int $gid The id of the group.
+     * @param bool $stay_leader Whether the old owner should stay as a leader.
+     * @return bool Whether the transfer was successful.
+     */
+
+    public function transfer_ownership(int $original_owner=0, int $new_owner=0, int $gid=0,bool $stay_leader=true): bool
     {
         global $mybb, $db;
+        if($original_owner == $new_owner)
+        {
+            return false;
+        }
         if(!$stay_leader)
         {
             $db->delete_query("socialgroup_leaders", "gid=$gid AND uid=$original_owner");
@@ -416,12 +487,18 @@ class socialgroupsuserhandler
         {
             $this->add_leader($gid, $new_owner);
         }
+        return true;
     }
 
-    public function can_groupcp($userid)
+    /**
+     * This determines if a user can access Group CP.
+     * @param int $userid The id of the user.
+     * @return array|bool true if global moderator. False if no groups. Array if not moderator, but leader.
+     */
+
+    public function can_groupcp(int $userid=0)
     {
         global $mybb, $db;
-        $userid = (int) $userid;
         if($mybb->usergroup['issupermod'] || $mybb->usergroup['cancp'])
         {
             return true;
@@ -432,6 +509,7 @@ class socialgroupsuserhandler
         {
             $moderators[$moderator['mid']] = $moderator['uid'];
         }
+        $db->free_result($query);
         if(in_array($userid, $moderators))
         {
             return true;
@@ -447,6 +525,49 @@ class socialgroupsuserhandler
         {
             $my_groups[] = $leader['gid'];
         }
+        $db->free_result($query);
         return $my_groups;
+    }
+
+    /**
+     * This generates HTML for who is viewing a particular group.
+     * Uses the who's online cutoff.
+     * @param int $gid The id of the group.
+     */
+
+    public function viewing_group(int $gid=0)
+    {
+        global $db, $mybb, $socialgroups, $templates, $online_members;
+        if(!$gid)
+        {
+            $socialgroups->error("invalid_group");
+        }
+        $cutoff = time() - 60 * $mybb->settings['wolcutoffmins'];
+
+        // Account for invisible users.
+        $invisible = 1;
+        if($mybb->usergroup['canviewwolinvis'])
+        {
+            $invisible = 0;
+        }
+
+        $query = $db->query("SELECT u.username, u.usergroup, u.displaygroup, u.invisible, s.uid FROM " . TABLE_PREFIX . "users u
+        LEFT JOIN " . TABLE_PREFIX . "sessions s ON(u.uid=s.uid)
+        WHERE s.location LIKE '%showgroup.php?gid=" . $gid
+        . "%' AND s.time >= " . $cutoff . " AND u.invisible >= " . $invisible);
+        $comma = "";
+        while($user = $db->fetch_array($query))
+        {
+            $invisiblemark = "";
+            if($user['invisible'] && $mybb->usergroup['canviewwolinvis'])
+            {
+                $invisiblemark = "*";
+            }
+            $user['formattedname'] = format_name($user['username'], $user['usergroup'], $user['displaygroup']);
+            $user['profilelink'] = build_profile_link($user['formattedname'], $user['uid']);
+            eval("\$online_members .= \"".$templates->get("forumdisplay_usersbrowsing_user")."\";");
+            $comma = ",";
+        }
+        $db->free_result($query);
     }
 }
